+++ main.go
@@ -58,7 +58,7 @@ import (
 )
 
 // ReleaseVersion is the release version for the code.
-var ReleaseVersion = "0.5.5"
+var ReleaseVersion = "0.5.6-dev"
 
 func main() {
 	os.Exit(main2())